feat: Personal summary is now in list format.
The personal summary section is now a list of points rather than a
paragraph to make it quicker to read and scan.

New styles have been added for the list element (.hero__list) and the
sectioning element (.hero__content) has been improved to
standardise margin and padding between each section of the hero block
(title, personal summary and actions) and therefore eliminating the need
for the .hero__actions and .hero__lead elements. As a result of these
changes the file hero.scss has been restructured.