fix: Migrate to Gatsby 3.x
Update Gatsby and related packages to latest version to reduce technical
debt and fix vulnerabilities. This is a semver major update but the site
is not impacted by any of the breaking changes.

Running the develop server (npm run develop) does throw a couple of
warnings which should be addressed in future commits.