fix: hero text is no longer blurry in firefox
Replaces a css transform based slant with a clip-path based slant to
make hero component content easier to read in firefox.

The trade-off of the clip-path approach is this technique is not
supported in older browsers such as MS Edge. In these browsers the
component gracefully degrades as the clip-path rule will be ignored and
the bottom edge will be a straight line instead of a slant.